[REF] forwardport: move PR update tests to their own module
